refactor: split editor runtime modules
