+++ doc/wiki/Home.md
@@ -4,14 +4,15 @@
 
 ## Architecture
 
-The project follows the same split used by common graphics tools:
+The runtime is organized around an application shell plus a layered workspace:
 
-- `Window` hosts the application shell and platform events
-- `Document` stores graphic objects and selection state
-- `CanvasView` controls zoom, pan, viewport, and coordinate transforms
-- `ToolController` routes user input into editing tools
-- `RenderSystem` draws document geometry and overlays
-- `UiSystem` builds the editor chrome (toolbar, inspector, status bar)
+- `application.c` owns bootstrap, GLFW callbacks, and frame orchestration
+- `Workspace` splits into `EditorCore`, `EditorSession`, and `EditorServices`
+- `Document` owns objects, layers, IDs, and spatial queries
+- `CommandExecutor` owns undo/redo and transactions
+- `ToolController` is registry-backed and focused on active-tool runtime
+- `UiSystem` is split into runtime, theme, layout, dialog, chrome, and inspector modules
+- Built-in objects and tools are registered through app-level manifests
 
 ## Design Goals
 
@@ -34,7 +35,10 @@ The project follows the same split used by common graphics tools:
 
 ## Where to Start
 
-1. [Getting Started](Getting-Started) — build and run the application
-2. [Data Flow](Data-Flow) — understand how events become drawings
-3. [Extending](Extending) — add new object types or tools
-4. [Controls](../../README.md#controls) — keyboard and mouse shortcuts
+1. [Getting Started](getting-started.md) — build and run the application
+2. [Architecture](architecture.md) — high-level runtime structure and current boundaries
+3. [Core Systems Overview](core-systems-overview.md) — responsibilities of document, commands, tools, UI, and render
+4. [Data Flow](data-flow.md) — understand how events become editor actions and renders
+5. [Extending](extending.md) — add new object types or tools
+6. [Layering Plan](architecture-layering-plan.md) — historical refactor roadmap and target boundaries
+7. [Controls](../../README.md#controls) — keyboard and mouse shortcuts
