+++ include/app/object_manifest.h
@@ -1,0 +1,8 @@
+#ifndef GLDRAW_APP_OBJECT_MANIFEST_H
+#define GLDRAW_APP_OBJECT_MANIFEST_H
+
+int object_manifest_register_builtins(void);
+int object_manifest_register_fake_star(void);
+int object_manifest_register_all(void);
+
+#endif /* GLDRAW_APP_OBJECT_MANIFEST_H */
