+++ include/app/tool_manifest.h
@@ -1,0 +1,6 @@
+#ifndef GLDRAW_APP_TOOL_MANIFEST_H
+#define GLDRAW_APP_TOOL_MANIFEST_H
+
+int tool_manifest_register_all(void);
+
+#endif /* GLDRAW_APP_TOOL_MANIFEST_H */
